src/requirements.txt
Django==1.10.3
django-cors-headers==2.1.0
djangorestframework==3.5.3
Faker==0.7.3
psycopg2==2.7.4